chore: split fetch model api call out to discrete fn
